+++ client/src/components/NavHeader.js
@@ -1,13 +1,17 @@
 import React from "react";
 import {Link} from 'react-router-dom';
+import Logo from '../static/Hello-World-Logo.png';
 
 
 function NavHeader(props) {
   
     return (
       <div className="App">
-        <Link to='/'><button>Home</button></Link>
+        <img src = {Logo} width = "50%"/>
+        <br />
+        <Link to='/home'><button>Home</button></Link>
         <Link to='/charts'><button>Charts</button></Link>
+        <Link to='/quiz'><button>Quiz</button></Link>
       </div>
     );
   }
