+++ client/src/components/PlanetCarousel.js
@@ -1,6 +1,6 @@
 import React, {useState} from 'react';
 
-const PlanetCarousel = ({ planets, slides }) => {
+const PlanetCarousel = ({ planets }) => {
 
     const [current, setCurrent] = useState(3)
     const length = planets.length
@@ -16,7 +16,6 @@ const PlanetCarousel = ({ planets, slides }) => {
     if (!Array.isArray(planets) || planets.length <= 0){
         return null;
     }
-
 
     return(
         <section>
