+++ client/src/static/CSS/SolarInfo.css
@@ -2,7 +2,7 @@
     display: flex;
     justify-content: start;
     padding-left: 25px;
-    background-image: url(../static/EarthAvatar.jpeg);
+    background-image: url(../EarthAvatar.jpeg);
     background-size: 15px;
     background-repeat: no-repeat;
 }
